requirements.txt
gpxpy
numpy
imageio
Pillow
imageio[ffmpeg]
PyOpenGL
PyOpenGL_accelerate
pygame